+++ picachain/datastore/__init__.py
@@ -1,5 +1,5 @@
+from picachain.datastore.base import DataStore
 from picachain.datastore.chroma import ChromaStore
-from picachain.datastore.datastore import DataStore
 from picachain.datastore.pinecone import PineconeStore
 
 __all__ = [
