+++ picachain/embedding/__init__.py
@@ -1,5 +1,5 @@
+from picachain.embedding.base import Embedding
 from picachain.embedding.clip_embedding import ClipEmbedding
-from picachain.embedding.embedding import Embedding
 
 __all__ = [
     "Embedding",
