+++ picachain/models/base.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+
+class Model(ABC):
+    """Model interface"""
